Add filters for branch, tag and PR to different classes for modularity
