+++ src/main/java/net/gleske/scmfilter/impl/trait/WildcardSCMBranchFilterTrait.java
@@ -1,0 +1,153 @@
+/*
+ * The MIT License
+ *
+ * Copyright (c) 2017, CloudBees, Inc.
+ * Copyright (c) 2017-2020, Sam Gleske - https://github.com/samrocketman
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in
+ * all copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+ * THE SOFTWARE.
+ */
+
+package net.gleske.scmfilter.impl.trait;
+
+import hudson.Extension;
+import java.util.regex.Pattern;
+import jenkins.scm.api.SCMHead;
+import jenkins.scm.api.SCMSource;
+import jenkins.scm.api.mixin.ChangeRequestSCMHead2;
+import jenkins.scm.api.mixin.ChangeRequestSCMHead;
+import jenkins.scm.api.mixin.TagSCMHead;
+import jenkins.scm.api.trait.SCMHeadPrefilter;
+import jenkins.scm.api.trait.SCMSourceContext;
+import jenkins.scm.api.trait.SCMSourceTrait;
+import jenkins.scm.api.trait.SCMSourceTraitDescriptor;
+import jenkins.scm.impl.trait.Selection;
+import org.apache.commons.lang.StringUtils;
+import org.jenkinsci.Symbol;
+import org.kohsuke.stapler.DataBoundConstructor;
+
+/**
+ * Decorates a {@link SCMSource} with a {@link SCMHeadPrefilter} that filters
+ * {@link SCMHead} instances based on matching wildcard include/exclude rules.
+ *
+ * @since 0.5
+ */
+public class WildcardSCMBranchFilterTrait extends SCMSourceTrait {
+
+    /**
+     * The branch include rules.
+     */
+    private final String branchIncludes;
+
+    /**
+     * The branch exclude rules.
+     */
+    private final String branchExcludes;
+
+    /**
+     * Stapler constructor.
+     *
+     * @param branchIncludes the branch include rules
+     * @param branchExcludes the branch exclude rules
+     */
+    @DataBoundConstructor
+    public WildcardSCMBranchFilterTrait(String branchIncludes, String branchExcludes) {
+        this.branchIncludes = StringUtils.defaultIfBlank(branchIncludes, "");
+        this.branchExcludes = StringUtils.defaultIfBlank(branchExcludes, "");
+    }
+
+    /**
+     * Returns the branch include rules.
+     *
+     * @return the branch include rules.
+     */
+    public String getBranchIncludes() {
+        return branchIncludes;
+    }
+
+    /**
+     * Returns the branch exclude rules.
+     *
+     * @return the branch exclude rules.
+     */
+    public String getBranchExcludes() {
+        return branchExcludes;
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    @Override
+    protected void decorateContext(SCMSourceContext<?, ?> context) {
+        context.withPrefilter(new SCMHeadPrefilter() {
+            @Override
+            public boolean isExcluded(SCMSource request, SCMHead head) {
+                if (!(head instanceof TagSCMHead || head instanceof ChangeRequestSCMHead
+                        || head instanceof ChangeRequestSCMHead2)) {
+                    // branch
+                    return !Pattern.matches(getPattern(getBranchIncludes()), head.getName())
+                            || Pattern.matches(getPattern(getBranchExcludes()), head.getName());
+                }
+
+                return false;
+            }
+        });
+    }
+
+    /**
+     * Returns the pattern corresponding to the branches containing wildcards.
+     *
+     * @param branches the names of branches to create a pattern for
+     * @return pattern corresponding to the branches containing wildcards
+     */
+    private String getPattern(String branches) {
+        StringBuilder quotedBranches = new StringBuilder();
+        for (String wildcard : branches.split(" ")) {
+            StringBuilder quotedBranch = new StringBuilder();
+            for (String branch : wildcard.split("(?=[*])|(?<=[*])")) {
+                if (branch.equals("*")) {
+                    quotedBranch.append(".*");
+                } else if (!branch.isEmpty()) {
+                    quotedBranch.append(Pattern.quote(branch));
+                }
+            }
+            if (quotedBranches.length() > 0) {
+                quotedBranches.append("|");
+            }
+            quotedBranches.append(quotedBranch);
+        }
+        return quotedBranches.toString();
+    }
+
+    /**
+     * Our descriptor.
+     */
+    @Symbol("WildcardSCMBranchFilter")
+    @Extension
+    @Selection
+    public static class DescriptorImpl extends SCMSourceTraitDescriptor {
+
+        /**
+         * {@inheritDoc}
+         */
+        @Override
+        public String getDisplayName() {
+            return Messages.WildcardSCMBranchFilterTrait_DisplayName();
+        }
+    }
+}
