+++ src/main/resources/net/gleske/scmfilter/impl/trait/Messages.properties
@@ -22,8 +22,11 @@
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 # THE SOFTWARE.
 #
-WildcardSCMFilterTrait.DisplayName=Filter by branch, tag or pull request (with wildcards)
 RegexSCMBranchFilterTrait.DisplayName=Filter by branch (with regular expression)
 RegexSCMTagFilterTrait.DisplayName=Filter by tag (with regular expression)
 RegexSCMPRDestinationFilterTrait.DisplayName=Filter by pull request destination branch (with regular expression)
 RegexSCMPROriginFilterTrait_DisplayName=Filter by pull request origin branch (with regular expression)
+WildcardSCMBranchFilterTrait.DisplayName=Filter by branch (with wildcards)
+WildcardSCMTagFilterTrait.DisplayName=Filter by tag (with wildcards)
+WildcardSCMPRDestinationFilter_DisplayName=Filter by pull request destination branch (with wildcards)
+WildcardSCMPROriginFilterTrait_DisplayName=Filter by pull request origin branch (with wildcards)
