--- src/main/resources/net/gleske/scmfilter/impl/trait/WildcardSCMFilterTrait/help-prDestinationExcludes.html
@@ -1,4 +1,0 @@
-<div>
-    Branch name patterns of pull requests destination to ignore even if matched by the includes list.
-    For example: <code>release</code>
-</div>
