--- src/main/resources/net/gleske/scmfilter/impl/trait/WildcardSCMFilterTrait/help-prOriginExcludes.html
@@ -1,4 +1,0 @@
-<div>
-    Branch name patterns of pull requests origin to ignore even if matched by the includes list.
-    For example: <code>release</code>
-</div>
