+++ src/main/resources/net/gleske/scmfilter/impl/trait/WildcardSCMPRDestinationFilterTrait/help-prDestinationExcludes.html
@@ -1,0 +1,4 @@
+<div>
+    Pull requests destination branch name patterns to ignore even if matched by the includes list.
+    For example: <code>release</code>
+</div>
