+++ src/main/resources/net/gleske/scmfilter/impl/trait/WildcardSCMPRDestinationFilterTrait/help-prDestinationIncludes.html
@@ -1,4 +1,4 @@
 <div>
-    Space-separated list of branch name patterns of pull requests origin to consider.
+    Space-separated list of pull request destination branch patterns to consider.
     You may use <code>*</code> as a wildcard; for example: <code>master release*</code>
 </div>
