+++ src/main/resources/net/gleske/scmfilter/impl/trait/WildcardSCMPROriginFilterTrait/help-prOriginExcludes.html
@@ -1,0 +1,4 @@
+<div>
+    Pull request origin branch name patterns to ignore even if matched by the includes list.
+    For example: <code>release</code>
+</div>
