+++ src/main/resources/net/gleske/scmfilter/impl/trait/WildcardSCMPROriginFilterTrait/help-prOriginIncludes.html
@@ -1,4 +1,4 @@
 <div>
-    Space-separated list of branch name patterns of pull requests destination to consider.
+    Space-separated list of pull request origin branch name patterns to consider.
     You may use <code>*</code> as a wildcard; for example: <code>master release*</code>
 </div>
